+++ src/option.c
@@ -177,4 +177,5 @@ void set_default_options(void)
    opt_set_int(OPT_MISSING_WAIT, 1);
    opt_set_int(OPT_LAYOUT_VERBOSE, get_int_env("NVC_LAYOUT_VERBOSE", 0));
    opt_set_int(OPT_SINGLE_UNIT, 0);
+   opt_set_int(OPT_PRESERVE_CASE, 0);
 }