+++ src/vhpi/vhpi-model.c
@@ -1120,11 +1120,10 @@ static void init_typeDecl(c_typeDecl *d, tree_t t, type_t type)
    char *pos = full, *lastpos;
    while (lastpos = pos, (pos = strchr(pos, '.')))
       *pos = ':';
-   d->decl.FullName = d->decl.FullCaseName = new_string(full);
+   d->decl.FullCaseName = new_string(full);
 
    // Kludge to set the name correctly for nested arrays
-   d->decl.Name = d->decl.CaseName =
-      (vhpiStringT)d->decl.FullName + (lastpos - full) + 1;
+   d->decl.CaseName = (vhpiStringT)d->decl.FullCaseName + (lastpos - full) + 1;
 
    d->type   = type;
    d->format = vhpi_format_for_type(d->type, &d->map_str);